// sdk/ts/src/generated/protos/virtengine/deployment/v1beta4/query.ts
export interface QueryDeploymentResponse {
  /** Deployment represents a deployment on the network. */
  deployment:
    | Deployment
    | undefined;
  /** Groups is a list of deployment groups. */
  groups: Group[];
  /**
   * EscrowAccount represents an escrow mechanism where funds are held.
   * This ensures that obligations of both tenants and providers involved in the
   * transaction are met without direct access to each other's accounts.
   */
  escrowAccount: Account | undefined;
}
// ...
export const QueryDeploymentResponse: MessageFns<
  QueryDeploymentResponse,
  "virtengine.deployment.v1beta4.QueryDeploymentResponse"
> = {
  $type: "virtengine.deployment.v1beta4.QueryDeploymentResponse" as const,

  encode(message: QueryDeploymentResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.deployment !== undefined) {
      Deployment.encode(message.deployment, writer.uint32(10).fork()).join();
    }
    for (const v of message.groups) {
      Group.encode(v!, writer.uint32(18).fork()).join();
    }
    if (message.escrowAccount !== undefined) {
      Account.encode(message.escrowAccount, writer.uint32(26).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): QueryDeploymentResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseQueryDeploymentResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.deployment = Deployment.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.groups.push(Group.decode(reader, reader.uint32()));
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.escrowAccount = Account.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): QueryDeploymentResponse {
    return {
      deployment: isSet(object.deployment) ? Deployment.fromJSON(object.deployment) : undefined,
      groups: globalThis.Array.isArray(object?.groups) ? object.groups.map((e: any) => Group.fromJSON(e)) : [],
      escrowAccount: isSet(object.escrow_account) ? Account.fromJSON(object.escrow_account) : undefined,
    };
  },

  toJSON(message: QueryDeploymentResponse): unknown {
    const obj: any = {};
    if (message.deployment !== undefined) {
      obj.deployment = Deployment.toJSON(message.deployment);
    }
    if (message.groups?.length) {
      obj.groups = message.groups.map((e) => Group.toJSON(e));
    }
    if (message.escrowAccount !== undefined) {
      obj.escrow_account = Account.toJSON(message.escrowAccount);
    }
    return obj;
  },

  create(base?: DeepPartial<QueryDeploymentResponse>): QueryDeploymentResponse {
    return QueryDeploymentResponse.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<QueryDeploymentResponse>): QueryDeploymentResponse {
    const message = createBaseQueryDeploymentResponse();
    message.deployment = (object.deployment !== undefined && object.deployment !== null)
      ? Deployment.fromPartial(object.deployment)
      : undefined;
    message.groups = object.groups?.map((e) => Group.fromPartial(e)) || [];
    message.escrowAccount = (object.escrowAccount !== undefined && object.escrowAccount !== null)
      ? Account.fromPartial(object.escrowAccount)
      : undefined;
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

type DeepPartial<T> = T extends Builtin ? T
  : T extends Long ? string | number | Long : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

interface MessageFns<T, V extends string> {
  readonly $type: V;
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create(base?: DeepPartial<T>): T;
  fromPartial(object: DeepPartial<T>): T;
}